+++ firmware/src/UI/Widgets/HeaderBar.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../Core/ScreenID.h"
+
+class HeaderBar
+{
+public:
+    static void Draw(ScreenID current);
+};