+++ firmware/src/UI/Widgets/TitleBar.h
@@ -1,7 +1,0 @@
-#pragma once
-
-class TitleBar
-{
-public:
-    static void Draw();
-};